refactor: Activity에서 Handler 추출
