+++ app/src/main/java/woowacourse/shopping/view/cart/CartActivity.kt
@@ -14,25 +14,44 @@ import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityCartBinding
 import woowacourse.shopping.view.cart.adatper.CartAdapter
 import woowacourse.shopping.view.cart.adatper.CartAdapterEventHandler
+import woowacourse.shopping.view.cart.event.CartScreenEventHandler
 import woowacourse.shopping.view.cart.vm.CartViewModel
 import woowacourse.shopping.view.cart.vm.CartViewModelFactory
 
-class CartActivity : AppCompatActivity(), CartAdapterEventHandler, CartScreenEventHandler {
+class CartActivity : AppCompatActivity() {
     private lateinit var binding: ActivityCartBinding
     private val viewModel: CartViewModel by viewModels { CartViewModelFactory() }
     private val cartAdapter by lazy {
         CartAdapter(
             items = viewModel.products.value ?: emptyList(),
-            handler = this,
+            handler = cartAdapterEvent,
         )
     }
+
+    private val cartAdapterEvent =
+        object : CartAdapterEventHandler {
+            override fun onClickDeleteItem(id: Long) {
+                viewModel.deleteProduct(id)
+            }
+        }
+
+    private val cartScreenEvent =
+        object : CartScreenEventHandler {
+            override fun onClickNextPage() {
+                viewModel.addPage()
+            }
+
+            override fun onClickPreviousPage() {
+                viewModel.subPage()
+            }
+        }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = DataBindingUtil.setContentView(this, R.layout.activity_cart)
         with(binding) {
             lifecycleOwner = this@CartActivity
-            eventHandler = this@CartActivity
+            eventHandler = cartScreenEvent
             adapter = cartAdapter
             vm = viewModel
         }
@@ -63,18 +82,6 @@ class CartActivity : AppCompatActivity(), CartAdapterEventHandler, CartScreenEve
         viewModel.products.observe(this) { value ->
             cartAdapter.submitList(value)
         }
-    }
-
-    override fun onClickDeleteItem(id: Long) {
-        viewModel.deleteProduct(id)
-    }
-
-    override fun onClickNextPage() {
-        viewModel.addPage()
-    }
-
-    override fun onClickPreviousPage() {
-        viewModel.subPage()
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
